// src copy/components/KycModal/messages.d.ts
export declare const messages: {
    overdueTitle: {
        id: string;
    };
    overdueMessage: {
        id: string;
    };
    dueDateLabel: {
        id: string;
    };
    kycTitle: {
        id: string;
    };
    dueDateWarningMessage: {
        id: string;
    };
    kycReasonMessage: {
        id: string;
    };
    creditConsentLabel: {
        id: string;
    };
    updateNowButton: {
        id: string;
    };
    remindLaterButton: {
        id: string;
    };
};